third_party/blink/web_tests/jquery/effects-expected.txt

ALERT: 420 tests of 420 passed, 0 failed.
